Polish Pass: Part 1 (#4864)

+++ Kavita.Services/BookmarkService.cs
@@ -185,7 +185,11 @@ public class BookmarkService(
             return false;
         }
 
-        await DeleteBookmarkFiles([bookmarkToDelete], ct);
+        if (bookmarkToDelete != null)
+        {
+            await DeleteBookmarkFiles([bookmarkToDelete], ct);
+        }
+
         return true;
     }
